added signing configs, androidx dependencies and fixed error of cmake
1-build.gradle
 - added keystore file and signing configs code in order to generate signed apk(s) of release build
 - removed publishNonDefault as it was deprecated
 - specifying which ndkversion we are using as per gradle warning
 - changed versioncode and versionname to release beta
 - implementation of androidx dependencies
2-keystore.properties, RohitKeyStore.jks
 - holds the access information of keystore
3-native-lib.cpp
 - used static_cast as per cmake error
4-gradle-wrapper.properties
 - upgrading gradle version to 6.3
5-Below files are edited in order to implement androidx
 - ExampleInstrumentedTest.kt
 - MainActivity.kt
 - SettingsActivity.kt
 - Wrap.kt
 - activity_main.xml
 - activity_settings.xml
 - content_main.xml and removed previous developer name as well
 - gradle.properties

+++ code/mobile/android/PhoneVR/app/src/main/cpp/native-lib.cpp
@@ -180,7 +180,7 @@ SUB(startMediaCodec)(JNIEnv *env, jclass, jobject surface) {
                 {
                     size_t bufSz = 0;
                     uint8_t *buf = AMediaCodec_getInputBuffer(codec, (size_t)idx, &bufSz);
-                    PVREnqueueVideoBuf({buf, idx, bufSz}); // into emptyVbuf
+                    PVREnqueueVideoBuf({buf, static_cast<int>(idx), bufSz}); // into emptyVbuf
                     PVR_DB("[MediaCodec th] Getting MCInputMedia Buf[ "+ to_string(bufSz)+ "] @ idx:" + to_string(idx) + ". into emptyVbuf");
                 }
             }
